simulado/requirements.txt
pandas
ipykernel
openpyxl
selenium
python-dotenv